refactor: reduce code duplication
