+++ fdm-app/app/routes/farm.$b_id_farm.$calendar.balance.organic-matter.$b_id.tsx
@@ -21,6 +21,7 @@ import {
   useLocation,
 } from "react-router"
 import { BufferStripWarning } from "~/components/blocks/balance/buffer-strip-warning"
+import { MissingParametersWarning } from "~/components/blocks/balance/missing-parameters-warning"
 import { OrganicMatterBalanceChart } from "~/components/blocks/balance/organic-matter-chart"
 import OrganicMatterBalanceDetails from "~/components/blocks/balance/organic-matter-details"
 import { NitrogenBalanceFallback } from "~/components/blocks/balance/skeletons"
@@ -145,30 +146,6 @@ export default function FarmBalanceOrganicMatterFieldBlock() {
   )
 }
 
-function MissingParametersErrorDisplay({ message }: { message: string }) {
-  const listItems = [
-    message.match(/a_n_rt/) ? <li key="a_n_rt">Totaal stikstofgehalte</li> : null,
-    message.match(/b_soiltype_agr/) ? <li key="b_soiltype_agr">Agrarisch bodemtype</li> : null,
-    message.match(/a_c_of|a_som_loi/) ? (
-      <li key="a_c_of_a_som_loi">Organische stofgehalte</li>
-    ) : null,
-  ].filter((item) => item != null)
-
-  return (
-    <div className="text-muted-foreground">
-      {listItems.length > 0 ? (
-        <>
-          <p>Voor dit perceel zijn de benodigde bodemparameters niet bekend:</p>
-          <br />
-          <ul className="list-inside list-disc">{...listItems}</ul>
-        </>
-      ) : (
-        <p>Voor dit perceel zijn enkele benodigde bodemparameters niet bekend.</p>
-      )}
-    </div>
-  )
-}
-
 function OrganicMatterBalance({
   farm,
   field,
@@ -205,7 +182,7 @@ function OrganicMatterBalance({
           </CardHeader>
           <CardContent>
             {fieldResult.errorMessage.match(/Missing required soil parameters/) ? (
-              <MissingParametersErrorDisplay message={fieldResult.errorMessage} />
+              <MissingParametersWarning message={fieldResult.errorMessage} />
             ) : (
               <div className="text-muted-foreground">
                 <p>
